refactor : Add directory structure, Add restart message
