+++ src/main/java/baseball/generator/ComputerNumberGenerator.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.generator;
 
 import camp.nextstep.edu.missionutils.Randoms;
