+++ src/main/java/baseball/generator/ResultGenerator.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.generator;
 
 import java.util.List;
