+++ src/main/java/baseball/view/GameMessage.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.view;
 
 public enum GameMessage {
     GAME_START_MESSAGE("숫자 야구 게임을 시작합니다.\n"),
