+++ src/test/java/baseball/InputValidatorTest.java
@@ -1,5 +1,6 @@
 package baseball;
 
+import baseball.validator.InputValidator;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
