+++ src/test/java/baseball/OutputViewViewTest.java
@@ -1,5 +1,6 @@
 package baseball;
 
+import baseball.view.OutputView;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
